requirements.txt
streamlit==1.28.0
requests==2.31.0
pandas==2.1.1
textblob==0.17.1
plotly==5.17.0
nltk==3.8.1
python-dotenv==1.0.0
